refactor: do not use orderTitle as key
